+++ packages/graphiql/resources/build.sh
@@ -9,6 +9,7 @@ if [ ! -d "node_modules/.bin" ]; then
 fi
 
 babel src --ignore __tests__ --out-dir dist/
+ESM=true babel src --ignore __tests__ --out-dir esm/
 echo "Bundling graphiql.js..."
 browserify -g browserify-shim -s GraphiQL dist/index.js > graphiql.js
 echo "Bundling graphiql.min.js..."