+++ packages/graphiql/src/components/QueryHistory.js
@@ -14,32 +14,26 @@ import HistoryQuery from './HistoryQuery';
 const MAX_QUERY_SIZE = 100000;
 const MAX_HISTORY_LENGTH = 20;
 
-const shouldSaveQuery = (nextProps, current, lastQuerySaved) => {
-  if (nextProps.queryID === current.queryID) {
-    return false;
-  }
+const shouldSaveQuery = (query, variables, lastQuerySaved) => {
   try {
-    parse(nextProps.query);
+    parse(query);
   } catch (e) {
     return false;
   }
   // Don't try to save giant queries
-  if (nextProps.query.length > MAX_QUERY_SIZE) {
+  if (query.length > MAX_QUERY_SIZE) {
     return false;
   }
   if (!lastQuerySaved) {
     return true;
   }
-  if (
-    JSON.stringify(nextProps.query) === JSON.stringify(lastQuerySaved.query)
-  ) {
+  if (JSON.stringify(query) === JSON.stringify(lastQuerySaved.query)) {
     if (
-      JSON.stringify(nextProps.variables) ===
-      JSON.stringify(lastQuerySaved.variables)
+      JSON.stringify(variables) === JSON.stringify(lastQuerySaved.variables)
     ) {
       return false;
     }
-    if (!nextProps.variables && !lastQuerySaved.variables) {
+    if (variables && !lastQuerySaved.variables) {
       return false;
     }
   }
@@ -58,7 +52,11 @@ export class QueryHistory extends React.Component {
 
   constructor(props) {
     super(props);
-    this.historyStore = new QueryStore('queries', props.storage, MAX_HISTORY_LENGTH);
+    this.historyStore = new QueryStore(
+      'queries',
+      props.storage,
+      MAX_HISTORY_LENGTH,
+    );
     // favorites are not automatically deleted, so there's no need for a max length
     this.favoriteStore = new QueryStore('favorites', props.storage, null);
     const historyQueries = this.historyStore.fetchAll();
@@ -67,33 +65,14 @@ export class QueryHistory extends React.Component {
     this.state = { queries };
   }
 
-  componentWillReceiveProps(nextProps) {
-    if (
-      shouldSaveQuery(nextProps, this.props, this.historyStore.fetchRecent())
-    ) {
-      const item = {
-        query: nextProps.query,
-        variables: nextProps.variables,
-        operationName: nextProps.operationName,
-      };
-      this.historyStore.push(item);
-      const historyQueries = this.historyStore.items;
-      const favoriteQueries = this.favoriteStore.items;
-      const queries = historyQueries.concat(favoriteQueries);
-      this.setState({
-        queries,
-      });
-    }
-  }
-
   render() {
     const queries = this.state.queries.slice().reverse();
-    const queryNodes = queries.map(query => {
+    const queryNodes = queries.map((query, i) => {
       return (
         <HistoryQuery
           handleEditLabel={this.editLabel}
           handleToggleFavorite={this.toggleFavorite}
-          key={query.query}
+          key={`${i}:${query.label || query.query}`}
           onSelect={this.props.onSelectQuery}
           {...query}
         />
@@ -110,6 +89,24 @@ export class QueryHistory extends React.Component {
     );
   }
 
+  // Public API
+  updateHistory = (query, variables, operationName) => {
+    if (shouldSaveQuery(query, variables, this.historyStore.fetchRecent())) {
+      this.historyStore.push({
+        query,
+        variables,
+        operationName,
+      });
+      const historyQueries = this.historyStore.items;
+      const favoriteQueries = this.favoriteStore.items;
+      const queries = historyQueries.concat(favoriteQueries);
+      this.setState({
+        queries,
+      });
+    }
+  };
+
+  // Public API
   toggleFavorite = (query, variables, operationName, label, favorite) => {
     const item = {
       query,
@@ -129,6 +126,7 @@ export class QueryHistory extends React.Component {
     });
   };
 
+  // Public API
   editLabel = (query, variables, operationName, label, favorite) => {
     const item = {
       query,